Fixes: CND infinite lives, F-Zero extra swaps
Chip and Dale 1 (NES): re-enabled and infinite lives now working. Simply forgot to specify which RAM when I changed how this works, oops
F-Zero (SNES): Now does not swap unless you are actively racing! This will prevent swaps on completion of a level as well as retries.
Super Mario Land (GB): getting this to show 69 lives when infinite lives is on is kind of a bother, because of when the display updates. But at least you don't get shuffled on collecting a life anymore!
A better explanation about using this with BizHawk 2.8 is given.
Huge thanks to Diabetus for a massive playthrough that pointed to these bugs to fix.