Demon's Crest (initial), SMK tweak, update readme re 2.8
Plugin: As Mega Man Damage Shuffler is now compatible with 2.8 (thanks kalimag) and does not clash with this plugin anymore due to the mainmemory bug, we'll plan to move entirely to 2.8 once enough testing is done.
Demon's Crest (SNES): initial support
Super Mario Kart: commented out "swap on shrunk by mushroom" for now, multiple folks prefer this not to swap!