Per-game grace period, testing with F-Zero
The plugin already allows for the global "grace period" before a new game can swap to be increased from 10 frames. Testing an option to modify that on a per-game basis, using F-Zero as an example because it notoriously combos players with swaps. The plugin goes with whatever's largest out of the game's grace period, the global grace period, or 10 frames (still the minimum).