Star Fox 64 (N64) fixes
Thanks to the decomp, got a lot more info on how game states work and how to determine which set of damage buffer/iframe addresses the game uses (space levels and planet levels). Since we're using the damage buffer function, we do want to shuffle on lives (but not double-shuffle), and a small grace period equivalent to the usual iframe time covers that.