Kabuki QF - stop wrong shuffles on game starting
A wrong swap was happening on loading the first level. The cause was a combination of maxhp being set too high that it caught a garbage value (255) on loading the game plus hp using an unsigned byte. This caused a "HP drop" that was delayed in triggering a swap because swap_exceptions would not allow a shuffle before the timer went over 0 (as in, once you start a level). Tracked down the actual max HP and made the signed byte fixes.